@@ -15,24 +15,22 @@
         }
     ],
     "require": {
-        "php": ">=7.0",
-        "symfony/http-foundation": "2.8.*|3.0.*|3.1.*|3.2.*|3.3.*|3.4.*|4.0.*",
-        "tightenco/collect": "~5.0",
-        "opis/closure": "^2.3",
-        "mpociot/pipeline": "^1.0",
-        "react/socket": "~0.4",
+        "php": ">=7.1",
+        "symfony/http-foundation": "^2.8 || ^3.0 || ^4.0 || ^5.0 || ^6.0",
+        "tightenco/collect": "^5.0 || ^6.0 || ^7.0 || ^8.0",
+        "opis/closure": "^2.3 || ^3.0",
+        "mpociot/pipeline": "^1.0.2",
+        "react/socket": "~1.0",
         "spatie/macroable": "^1.0",
         "psr/container": "^1.0"
     },
     "require-dev": {
         "codeigniter/framework": "~3.0",
-        "illuminate/support": "^5.4",
-        "orchestra/testbench": "~3.0",
-        "phpunit/phpunit": "~6.4",
-        "mockery/mockery": "dev-master",
-        "sllh/php-cs-fixer-styleci-bridge": "^2.1",
+        "orchestra/testbench": "^3.4 || ^4.0 || ^5.0 || ^6.0",
+        "phpunit/phpunit": "^6.0 || ^7.0 || ^8.0",
+        "mockery/mockery": "^1.1",
         "doctrine/cache": "^1.6",
-        "symfony/cache": "^3.1",
+        "symfony/cache": "^3.4.35 || ^4.3.7 || ^5.0 || ^6.0",
         "ext-curl": "*"
     },
     "suggest": {
@@ -51,7 +49,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "3.0-dev"
+            "dev-master": "2.0"
         },
         "laravel": {
             "providers": [
@@ -63,7 +61,8 @@
         }
     },
     "scripts": {
-        "test": "phpunit",
-        "cs": "php-cs-fixer fix"
-    }
+        "test": "phpunit"
+    },
+    "prefer-stable": true,
+    "minimum-stability": "dev"
 }
