@@ -13,10 +13,13 @@ class ApiAi implements MiddlewareInterface
     /** @var string */
     protected $token;
 
+    /** @var string */
+    protected $lang = 'en';
+
     /** @var HttpInterface */
     protected $http;
 
-    /** @var stdClass */
+    /** @var \stdClass */
     protected $response;
 
     /** @var string */
@@ -29,24 +32,27 @@ class ApiAi implements MiddlewareInterface
     protected $listenForAction = false;
 
     /**
-     * API.ai constructor.
+     * Wit constructor.
      * @param string $token api.ai access token
+     * @param string $lang language
      * @param HttpInterface $http
      */
-    public function __construct($token, HttpInterface $http)
+    public function __construct($token, HttpInterface $http, $lang = 'en')
     {
         $this->token = $token;
+        $this->lang = $lang;
         $this->http = $http;
     }
 
     /**
      * Create a new API.ai middleware instance.
      * @param string $token api.ai access token
+     * @param string $lang language
      * @return ApiAi
      */
-    public static function create($token)
+    public static function create($token, $lang = 'en')
     {
-        return new static($token, new Curl());
+        return new static($token, new Curl(), $lang);
     }
 
     /**
@@ -64,14 +70,14 @@ class ApiAi implements MiddlewareInterface
     /**
      * Perform the API.ai API call and cache it for the message.
      * @param  \BotMan\BotMan\Messages\Incoming\IncomingMessage $message
-     * @return stdClass
+     * @return \stdClass
      */
     protected function getResponse(IncomingMessage $message)
     {
         $response = $this->http->post($this->apiUrl, [], [
             'query' => [$message->getText()],
             'sessionId' => md5($message->getConversationIdentifier()),
-            'lang' => 'en',
+            'lang' => $this->lang,
         ], [
             'Authorization: Bearer '.$this->token,
             'Content-Type: application/json; charset=utf-8',
@@ -109,17 +115,28 @@ class ApiAi implements MiddlewareInterface
     {
         $response = $this->getResponse($message);
 
-        $reply = isset($response->result->fulfillment->speech) ? $response->result->fulfillment->speech : '';
-        $action = isset($response->result->action) ? $response->result->action : '';
+        $reply = $response->result->fulfillment->speech ?? '';
+        $action = $response->result->action ?? '';
         $actionIncomplete = isset($response->result->actionIncomplete) ? (bool) $response->result->actionIncomplete : false;
-        $intent = isset($response->result->metadata->intentName) ? $response->result->metadata->intentName : '';
+        $intent = $response->result->metadata->intentName ?? '';
         $parameters = isset($response->result->parameters) ? (array) $response->result->parameters : [];
+        $responseMessages = isset($response->result->fulfillment->messages) ? json_decode(json_encode($response->result->fulfillment->messages), true) : [];
+        $contexts = isset($response->result->contexts) ? json_decode(json_encode($response->result->contexts), true) : [];
+
+        if (count($responseMessages)) {
+            $textResponses = count(collect($responseMessages)->where('type', 0)) ? array_values(collect($responseMessages)->where('type', 0)->toArray()) : [];
+            $customPayloadResponses = count(collect($responseMessages)->where('type', 4)) ? array_values(collect($responseMessages)->where('type', 4)->toArray()) : [];
+        }
 
         $message->addExtras('apiReply', $reply);
         $message->addExtras('apiAction', $action);
         $message->addExtras('apiActionIncomplete', $actionIncomplete);
         $message->addExtras('apiIntent', $intent);
         $message->addExtras('apiParameters', $parameters);
+        $message->addExtras('apiResponseMessages', $responseMessages);
+        $message->addExtras('apiTextResponses', $textResponses ?? []);
+        $message->addExtras('apiCustomPayloadResponses', $customPayloadResponses ?? []);
+        $message->addExtras('apiContexts', $contexts);
 
         return $next($message);
     }
