@@ -3,9 +3,9 @@
 namespace BotMan\BotMan;
 
 use BotMan\BotMan\Cache\LaravelCache;
-use Illuminate\Support\ServiceProvider;
 use BotMan\BotMan\Container\LaravelContainer;
 use BotMan\BotMan\Storages\Drivers\FileStorage;
+use Illuminate\Support\ServiceProvider;
 
 class BotManServiceProvider extends ServiceProvider
 {
