@@ -2,8 +2,8 @@
 
 namespace BotMan\BotMan\Cache;
 
-use Psr\Cache\CacheItemPoolInterface;
 use BotMan\BotMan\Interfaces\CacheInterface;
+use Psr\Cache\CacheItemPoolInterface;
 
 class Psr6Cache implements CacheInterface
 {
