@@ -2,18 +2,19 @@
 
 namespace BotMan\BotMan\Commands;
 
-use Illuminate\Support\Collection;
-use BotMan\BotMan\Messages\Matcher;
-use BotMan\BotMan\Messages\Incoming\Answer;
-use BotMan\BotMan\Messages\Attachments\File;
 use BotMan\BotMan\Interfaces\DriverInterface;
 use BotMan\BotMan\Messages\Attachments\Audio;
+use BotMan\BotMan\Messages\Attachments\Contact;
+use BotMan\BotMan\Messages\Attachments\File;
 use BotMan\BotMan\Messages\Attachments\Image;
-use BotMan\BotMan\Messages\Attachments\Video;
-use BotMan\BotMan\Middleware\MiddlewareManager;
 use BotMan\BotMan\Messages\Attachments\Location;
+use BotMan\BotMan\Messages\Attachments\Video;
+use BotMan\BotMan\Messages\Incoming\Answer;
 use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use BotMan\BotMan\Messages\Matcher;
 use BotMan\BotMan\Messages\Matching\MatchingMessage;
+use BotMan\BotMan\Middleware\MiddlewareManager;
+use Illuminate\Support\Collection;
 
 class ConversationManager
 {
@@ -48,6 +49,8 @@ class ConversationManager
             $parameters[] = $message->getAudio();
         } elseif ($messageText === Location::PATTERN) {
             $parameters[] = $message->getLocation();
+        } elseif ($messageText === Contact::PATTERN) {
+            $parameters[] = $message->getContact();
         } elseif ($messageText === File::PATTERN) {
             $parameters[] = $message->getFiles();
         }
@@ -63,7 +66,7 @@ class ConversationManager
      * @param bool $withReceivedMiddleware
      * @return array|MatchingMessage[]
      */
-    public function getMatchingMessages($messages, MiddlewareManager $middleware, Answer $answer, DriverInterface $driver, $withReceivedMiddleware = true) : array
+    public function getMatchingMessages($messages, MiddlewareManager $middleware, Answer $answer, DriverInterface $driver, $withReceivedMiddleware = true): array
     {
         $matcher = new Matcher();
         $messages = Collection::make($messages)->reject(function (IncomingMessage $message) {
