@@ -3,10 +3,10 @@
 namespace BotMan\BotMan\Drivers;
 
 use BotMan\BotMan\Http\Curl;
-use Illuminate\Support\Collection;
-use BotMan\BotMan\Interfaces\HttpInterface;
 use BotMan\BotMan\Interfaces\DriverInterface;
+use BotMan\BotMan\Interfaces\HttpInterface;
 use BotMan\BotMan\Interfaces\VerifiesService;
+use Illuminate\Support\Collection;
 use Symfony\Component\HttpFoundation\Request;
 
 class DriverManager
@@ -158,6 +158,7 @@ class DriverManager
                 return true;
             }
         }
+
         return false;
     }
 
