@@ -2,11 +2,11 @@
 
 namespace BotMan\BotMan\Drivers\Tests;
 
-use BotMan\BotMan\Http\Curl;
 use BotMan\BotMan\Drivers\NullDriver;
+use BotMan\BotMan\Http\Curl;
 use BotMan\BotMan\Interfaces\DriverInterface;
-use Symfony\Component\HttpFoundation\Request;
 use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * A driver that acts as a proxy for a global driver instance. Useful for mock/fake drivers in integration tests.
@@ -66,6 +66,11 @@ final class ProxyDriver implements DriverInterface
     }
 
     public function getUser(IncomingMessage $matchingMessage)
+    {
+        return self::instance()->getUser($matchingMessage);
+    }
+
+    public function getUserWithFields(array $fields, IncomingMessage $matchingMessage)
     {
         return self::instance()->getUser($matchingMessage);
     }
