@@ -53,7 +53,7 @@ class Curl implements HttpInterface
      * @param  array $headers [description]
      * @return resource
      */
-    protected static function prepareRequest($url, $parameters = [], $headers = [])
+    protected function prepareRequest($url, $parameters = [], $headers = [])
     {
         $request = curl_init();
 
