@@ -2,8 +2,9 @@
 
 namespace BotMan\BotMan\Messages\Incoming;
 
-use Illuminate\Support\Collection;
+use BotMan\BotMan\Messages\Attachments\Contact;
 use BotMan\BotMan\Messages\Attachments\Location;
+use Illuminate\Support\Collection;
 
 class IncomingMessage
 {
@@ -15,6 +16,9 @@ class IncomingMessage
 
     /** @var string */
     protected $recipient;
+
+    /** @var string */
+    protected $bot_id;
 
     /** @var array */
     protected $images = [];
@@ -37,15 +41,19 @@ class IncomingMessage
     /** @var \BotMan\BotMan\Messages\Attachments\Location */
     private $location;
 
+    /** @var \BotMan\BotMan\Messages\Attachments\Contact */
+    private $contact;
+
     /** @var bool */
     protected $isFromBot = false;
 
-    public function __construct($message, $sender, $recipient, $payload = null)
+    public function __construct($message, $sender, $recipient, $payload = null, $bot_id = '')
     {
         $this->message = $message;
         $this->sender = $sender;
         $this->recipient = $recipient;
         $this->payload = $payload;
+        $this->bot_id = $bot_id;
     }
 
     /**
@@ -85,7 +93,7 @@ class IncomingMessage
      */
     public function getConversationIdentifier()
     {
-        return 'conversation-'.sha1($this->getSender()).'-'.sha1($this->getRecipient());
+        return 'conversation-'.$this->bot_id.sha1($this->getSender()).'-'.sha1($this->getRecipient());
     }
 
     /**
@@ -95,7 +103,7 @@ class IncomingMessage
      */
     public function getOriginatedConversationIdentifier()
     {
-        return 'conversation-'.sha1($this->getSender()).'-'.sha1('');
+        return 'conversation-'.$this->bot_id.sha1($this->getSender()).'-'.sha1('');
     }
 
     /**
@@ -112,7 +120,7 @@ class IncomingMessage
 
     /**
      * @param string|null $key
-     * @return array
+     * @return mixed
      */
     public function getExtras($key = null)
     {
@@ -201,9 +209,33 @@ class IncomingMessage
     /**
      * @return \BotMan\BotMan\Messages\Attachments\Location
      */
-    public function getLocation() : Location
+    public function getLocation(): Location
     {
+        if (empty($this->location)) {
+            throw new \UnexpectedValueException('This message does not contain a location');
+        }
+
         return $this->location;
+    }
+
+    /**
+     * @return \BotMan\BotMan\Messages\Attachments\Contact
+     */
+    public function getContact(): Contact
+    {
+        if (empty($this->contact)) {
+            throw new \UnexpectedValueException('This message does not contain a contact');
+        }
+
+        return $this->contact;
+    }
+
+    /**
+     * @param \BotMan\BotMan\Messages\Attachments\Contact $contact
+     */
+    public function setContact(Contact $contact)
+    {
+        $this->contact = $contact;
     }
 
     /**
