@@ -2,8 +2,8 @@
 
 namespace BotMan\BotMan\Storages\Drivers;
 
-use Illuminate\Support\Collection;
 use BotMan\BotMan\Interfaces\StorageInterface;
+use Illuminate\Support\Collection;
 
 class FileStorage implements StorageInterface
 {
