@@ -2,21 +2,22 @@
 
 namespace BotMan\BotMan\tests;
 
-use Mockery as m;
 use BotMan\BotMan\BotMan;
-use PHPUnit\Framework\TestCase;
 use BotMan\BotMan\BotManFactory;
 use BotMan\BotMan\Cache\ArrayCache;
 use BotMan\BotMan\Drivers\DriverManager;
 use BotMan\BotMan\Drivers\Tests\FakeDriver;
 use BotMan\BotMan\Drivers\Tests\ProxyDriver;
 use BotMan\BotMan\Messages\Attachments\Audio;
+use BotMan\BotMan\Messages\Attachments\Contact;
 use BotMan\BotMan\Messages\Attachments\Image;
-use BotMan\BotMan\Messages\Attachments\Video;
 use BotMan\BotMan\Messages\Attachments\Location;
-use BotMan\BotMan\Tests\Fixtures\TestConversation;
+use BotMan\BotMan\Messages\Attachments\Video;
 use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use BotMan\BotMan\Tests\Fixtures\TestConversation;
 use BotMan\BotMan\Tests\Fixtures\TestDataConversation;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
 
 class BotManConversationTest extends TestCase
 {
@@ -29,17 +30,17 @@ class BotManConversationTest extends TestCase
     /** @var m\MockInterface */
     private $cache;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         DriverManager::loadDriver(ProxyDriver::class);
     }
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
         DriverManager::unloadDriver(ProxyDriver::class);
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->fakeDriver = new FakeDriver();
         $this->cache = m::mock(ArrayCache::class)->makePartial();
@@ -47,7 +48,7 @@ class BotManConversationTest extends TestCase
         $this->botman = BotManFactory::create([], $this->cache);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         ProxyDriver::setInstance(FakeDriver::createInactive());
         m::close();
@@ -320,19 +321,70 @@ class BotManConversationTest extends TestCase
         $this->botman->hears('Hello', function (BotMan $bot) {
             $bot->startConversation(new TestDataConversation());
         });
-
         $this->replyWithFakeMessage('Hello');
         static::assertEquals('What do you want to test?', $this->fakeDriver->getBotMessages()[0]->getText());
-
         $this->replyWithFakeMessage('location');
         static::assertEquals('Please supply a location', $this->fakeDriver->getBotMessages()[1]->getText());
-
         $message = new IncomingMessage(Location::PATTERN, 'helloman', '#helloworld');
         $location = new Location(41.123, -12.123);
         $message->setLocation($location);
         $this->replyWithFakeMessage($message);
-
         static::assertEquals('41.123:-12.123', $this->fakeDriver->getBotMessages()[2]->getText());
+    }
+
+    /** @test */
+    public function it_repeats_invalid_contact_answers()
+    {
+        $this->botman->hears('Hello', function (BotMan $bot) {
+            $bot->startConversation(new TestDataConversation());
+        });
+
+        $this->replyWithFakeMessage('Hello');
+        static::assertEquals('What do you want to test?', $this->fakeDriver->getBotMessages()[0]->getText());
+
+        $this->replyWithFakeMessage('contact');
+        static::assertEquals('Please supply a contact', $this->fakeDriver->getBotMessages()[1]->getText());
+
+        $this->replyWithFakeMessage('not_a_contact');
+        static::assertEquals('Please supply a contact', $this->fakeDriver->getBotMessages()[2]->getText());
+    }
+
+    /** @test */
+    public function it_calls_custom_repeat_method_on_invalid_contact_answers()
+    {
+        $this->botman->hears('Hello', function (BotMan $bot) {
+            $bot->startConversation(new TestDataConversation());
+        });
+
+        $this->replyWithFakeMessage('Hello');
+        static::assertEquals('What do you want to test?', $this->fakeDriver->getBotMessages()[0]->getText());
+
+        $this->replyWithFakeMessage('custom_contact_repeat');
+        static::assertEquals('Please supply a contact', $this->fakeDriver->getBotMessages()[1]->getText());
+
+        $this->replyWithFakeMessage('not_a_contact');
+        static::assertEquals('That is not a contact...', $this->fakeDriver->getBotMessages()[2]->getText());
+    }
+
+    /** @test */
+    public function it_returns_the_contact()
+    {
+        $this->botman->hears('Hello', function (BotMan $bot) {
+            $bot->startConversation(new TestDataConversation());
+        });
+
+        $this->replyWithFakeMessage('Hello');
+        static::assertEquals('What do you want to test?', $this->fakeDriver->getBotMessages()[0]->getText());
+
+        $this->replyWithFakeMessage('contact');
+        static::assertEquals('Please supply a contact', $this->fakeDriver->getBotMessages()[1]->getText());
+
+        $message = new IncomingMessage(Contact::PATTERN, 'helloman', '#helloworld');
+        $contact = new Contact('0775269856', 'Daniele', 'Rapisarda', '123');
+        $message->setcontact($contact);
+        $this->replyWithFakeMessage($message);
+
+        static::assertEquals('0775269856:Daniele:Rapisarda:123', $this->fakeDriver->getBotMessages()[2]->getText());
     }
 
     private function replyWithFakeMessage($message, $username = 'helloman', $channel = '#helloworld')
