@@ -3,12 +3,12 @@
 namespace BotMan\BotMan\Tests;
 
 use BotMan\BotMan\BotMan;
-use PHPUnit\Framework\TestCase;
 use BotMan\BotMan\BotManFactory;
 use BotMan\BotMan\Drivers\DriverManager;
 use BotMan\BotMan\Drivers\Tests\FakeDriver;
 use BotMan\BotMan\Drivers\Tests\ProxyDriver;
 use BotMan\BotMan\Interfaces\DriverEventInterface;
+use PHPUnit\Framework\TestCase;
 
 class BotManDriverEventTest extends TestCase
 {
@@ -17,24 +17,24 @@ class BotManDriverEventTest extends TestCase
     /** @var FakeDriver */
     private $fakeDriver;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         DriverManager::loadDriver(ProxyDriver::class);
     }
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
         DriverManager::unloadDriver(ProxyDriver::class);
     }
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->fakeDriver = new FakeDriver();
         ProxyDriver::setInstance($this->fakeDriver);
         $this->botman = BotManFactory::create([]);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         ProxyDriver::setInstance(FakeDriver::createInactive());
     }
@@ -50,6 +50,26 @@ class BotManDriverEventTest extends TestCase
         $this->botman->listen();
 
         $this->assertTrue($called);
+    }
+
+    /** @test */
+    public function it_can_assign_same_callback_for_events()
+    {
+        $called = 0;
+        $this->fakeDriver->hasMatchingEvent = new TestEvent([]);
+        $this->botman->on(['test_event', 'test_event_two'], function ($data, BotMan $bot) use (&$called) {
+            $called++;
+        });
+
+        $this->botman->listen();
+
+        $this->assertEquals(1, $called);
+
+        $this->fakeDriver->hasMatchingEvent = new TestEventTwo([]);
+
+        $this->botman->listen();
+
+        $this->assertEquals(2, $called);
     }
 
     /** @test */
@@ -111,6 +131,19 @@ class TestEvent implements DriverEventInterface
             'event' => 'test_event',
             'data' => 'foo',
         ];
+    }
+}
+
+class TestEventTwo extends TestEvent
+{
+    /**
+     * Return the event name to match.
+     *
+     * @return string
+     */
+    public function getName()
+    {
+        return 'test_event_two';
     }
 }
 
