@@ -2,39 +2,41 @@
 
 namespace BotMan\BotMan\tests;
 
-use Mockery as m;
 use BotMan\BotMan\BotMan;
-use Mockery\MockInterface;
-use PHPUnit\Framework\TestCase;
 use BotMan\BotMan\BotManFactory;
-use Illuminate\Support\Collection;
 use BotMan\BotMan\Cache\ArrayCache;
-use BotMan\BotMan\Drivers\NullDriver;
-use Psr\Container\ContainerInterface;
 use BotMan\BotMan\Drivers\DriverManager;
+use BotMan\BotMan\Drivers\NullDriver;
 use BotMan\BotMan\Drivers\Tests\FakeDriver;
-use BotMan\BotMan\Interfaces\UserInterface;
-use BotMan\BotMan\Messages\Incoming\Answer;
-use BotMan\BotMan\Tests\Fixtures\TestClass;
-use BotMan\BotMan\Tests\Fixtures\TestDriver;
-use BotMan\BotMan\Messages\Attachments\Audio;
-use BotMan\BotMan\Messages\Attachments\Image;
-use BotMan\BotMan\Messages\Attachments\Video;
-use Psr\Container\NotFoundExceptionInterface;
-use BotMan\BotMan\Tests\Fixtures\TestFallback;
-use BotMan\BotMan\Middleware\MiddlewareManager;
-use BotMan\BotMan\Messages\Attachments\Location;
-use BotMan\BotMan\Tests\Fixtures\TestMiddleware;
 use BotMan\BotMan\Exceptions\Base\BotManException;
-use BotMan\BotMan\Tests\Fixtures\TestConversation;
-use BotMan\BotMan\Messages\Incoming\IncomingMessage;
-use BotMan\BotMan\Tests\Fixtures\Middleware\Matching;
-use BotMan\BotMan\Tests\Fixtures\TestMatchMiddleware;
-use BotMan\BotMan\Messages\Conversations\Conversation;
-use BotMan\BotMan\Tests\Fixtures\TestAdditionalDriver;
-use BotMan\BotMan\Tests\Fixtures\TestNoMatchMiddleware;
 use BotMan\BotMan\Exceptions\Core\BadMethodCallException;
 use BotMan\BotMan\Exceptions\Core\UnexpectedValueException;
+use BotMan\BotMan\Interfaces\UserInterface;
+use BotMan\BotMan\Messages\Attachments\Audio;
+use BotMan\BotMan\Messages\Attachments\Contact;
+use BotMan\BotMan\Messages\Attachments\Image;
+use BotMan\BotMan\Messages\Attachments\Location;
+use BotMan\BotMan\Messages\Attachments\Video;
+use BotMan\BotMan\Messages\Conversations\Conversation;
+use BotMan\BotMan\Messages\Incoming\Answer;
+use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use BotMan\BotMan\Middleware\MiddlewareManager;
+use BotMan\BotMan\Tests\Fixtures\InvokableService;
+use BotMan\BotMan\Tests\Fixtures\Middleware\Matching;
+use BotMan\BotMan\Tests\Fixtures\TestAdditionalDriver;
+use BotMan\BotMan\Tests\Fixtures\TestClass;
+use BotMan\BotMan\Tests\Fixtures\TestConversation;
+use BotMan\BotMan\Tests\Fixtures\TestDriver;
+use BotMan\BotMan\Tests\Fixtures\TestFallback;
+use BotMan\BotMan\Tests\Fixtures\TestMatchMiddleware;
+use BotMan\BotMan\Tests\Fixtures\TestMiddleware;
+use BotMan\BotMan\Tests\Fixtures\TestNoMatchMiddleware;
+use Illuminate\Support\Collection;
+use Mockery as m;
+use Mockery\MockInterface;
+use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
+use Psr\Container\NotFoundExceptionInterface;
 
 /**
  * Class BotManTest.
@@ -49,12 +51,12 @@ class BotManTest extends TestCase
     /** @var ArrayCache */
     protected $cache;
 
-    public function tearDown()
+    protected function tearDown(): void
     {
         m::close();
     }
 
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->cache = new ArrayCache();
@@ -69,7 +71,16 @@ class BotManTest extends TestCase
         $driver = m::mock(FakeDriver::class)->makePartial();
 
         $driver->isBot = $data->get('is_from_bot', false);
-        $driver->messages = [new IncomingMessage($data->get('message'), $data->get('sender'), $data->get('recipient'))];
+
+        $messages = $data->get('message');
+
+        if (! is_array($messages)) {
+            $messages = [$messages];
+        }
+
+        $driver->messages = array_map(function ($message) use ($data) {
+            return new IncomingMessage($message, $data->get('sender'), $data->get('recipient'));
+        }, $messages);
 
         $botman->setDriver($driver);
 
@@ -204,7 +215,8 @@ class BotManTest extends TestCase
     /** @test */
     public function it_hears_matching_commands()
     {
-        $called = false;
+        $called_once = false;
+        $called_twice = false;
 
         $botman = $this->getBot([
             'sender' => 'UX12345',
@@ -212,11 +224,84 @@ class BotManTest extends TestCase
             'message' => 'Foo',
         ]);
 
-        $botman->hears('Foo', function ($bot) use (&$called) {
-            $called = true;
+        $botman->hears('Foo', function ($bot) use (&$called_once) {
+            $called_once = true;
         });
+
+        $botman->hears('Foo(.*)', function ($bot) use (&$called_twice) {
+            $called_twice = true;
+        });
+
         $botman->listen();
-        $this->assertTrue($called);
+        $this->assertTrue($called_once);
+        $this->assertTrue($called_twice);
+    }
+
+    /** @test */
+    public function it_hears_matching_commands_from_array()
+    {
+        $called = 0;
+
+        $botman = $this->getBot([
+            'sender' => 'UX12345',
+            'recipient' => 'general',
+            'message' => ['Foo', 'Bar'],
+        ]);
+
+        $botman->hears(['Foo', 'Bar'], function ($bot) use (&$called) {
+            $called++;
+        });
+
+        $botman->listen();
+
+        $this->assertEquals(2, $called);
+    }
+
+    /** @test */
+    public function it_resets_the_group_matching_with_multiple_commands()
+    {
+        $called = 0;
+
+        $botman = $this->getBot([
+            'sender' => 'UX12345',
+            'recipient' => 'general',
+            'message' => ['Foo 1', 'Bar 2'],
+        ]);
+
+        $botman->hears(['Foo (\d+)', 'Bar (\d+)'], function ($bot, $number) use (&$called) {
+            $called++;
+
+            $this->assertEquals($called, $number);
+        });
+
+        $botman->listen();
+    }
+
+    /** @test */
+    public function it_hears_only_first_matching_command_that_returns()
+    {
+        $called_once = false;
+        $called_twice = false;
+
+        $botman = $this->getBot([
+            'sender' => 'UX12345',
+            'recipient' => 'general',
+            'message' => 'Foo',
+        ]);
+
+        $botman->hears('Foo', function ($bot) use (&$called_once) {
+            $called_once = true;
+
+            return true;
+        });
+
+        $botman->hears('Foo(.*)', function ($bot) use (&$called_twice) {
+            $called_twice = true;
+        });
+
+        $botman->listen();
+        $this->assertTrue($called_once);
+        $this->assertFalse($called_twice);
     }
 
     /** @test */
@@ -367,6 +452,20 @@ class BotManTest extends TestCase
         ]);
         $botman->hears('foo', \stdClass::class);
         $botman->listen();
+    }
+
+    /** @test */
+    public function it_hears_matching_commands_without_invokable_service()
+    {
+        $botman = $this->getBot([
+            'sender' => 'UX12345',
+            'recipient' => 'general',
+            'message' => 'Foo',
+        ]);
+        $service = new InvokableService();
+        $botman->hears('foo', $service);
+        $botman->listen();
+        $this->assertTrue($service->invoked);
     }
 
     /** @test */
@@ -827,7 +926,7 @@ class BotManTest extends TestCase
             $conversation = new TestConversation();
 
             $botman->storeConversation($conversation, function (Answer $answer) use (&$called) {
-                $this->_throwException('called conversation');
+                throw new \Exception('called conversation');
             });
         });
         $botman->listen();
@@ -2280,6 +2379,39 @@ class BotManTest extends TestCase
             $called = true;
             $this->assertInstanceOf(Location::class, $data);
             $this->assertSame($location, $data);
+        });
+        $botman->listen();
+        $this->assertTrue($called);
+    }
+
+    /** @test */
+    public function it_returns_contact_as_second_argument()
+    {
+        $called = false;
+        $phone_number = '0775269856';
+        $first_name = 'Daniele';
+        $last_name = 'Rapisarda';
+        $user_id = '123';
+
+        $contact = new Contact($phone_number, $first_name, $last_name, $user_id);
+
+        $message = new IncomingMessage(Contact::PATTERN, '', '');
+        $message->setContact($contact);
+
+        $botman = $this->getBot([]);
+
+        $driver = m::mock(FakeDriver::class)->makePartial();
+        $driver->shouldReceive('getMessages')
+            ->andReturn([
+                $message,
+            ]);
+
+        $botman->setDriver($driver);
+
+        $botman->receivesContact(function ($bot, $data) use (&$called, $contact) {
+            $called = true;
+            $this->assertInstanceOf(Contact::class, $data);
+            $this->assertSame($contact, $data);
         });
         $botman->listen();
         $this->assertTrue($called);
