@@ -1,9 +1,9 @@
 <?php
 
-namespace BotMan\BotMan\Tests;
+namespace BotMan\BotMan\Tests\Cache;
 
-use PHPUnit\Framework\TestCase;
 use BotMan\BotMan\Cache\ArrayCache;
+use PHPUnit\Framework\TestCase;
 
 class ArrayCacheTest extends TestCase
 {
