@@ -1,17 +1,17 @@
 <?php
 
-namespace BotMan\BotMan\Tests;
+namespace BotMan\BotMan\Tests\Cache;
 
+use BotMan\BotMan\Cache\CodeIgniterCache;
 use CI_Cache;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
-use BotMan\BotMan\Cache\CodeIgniterCache;
 
 class CodeIgniterCacheTest extends TestCase
 {
     use \Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 
-    public function tearDown()
+    protected function tearDown(): void
     {
         m::close();
     }
