@@ -2,17 +2,17 @@
 
 namespace BotMan\BotMan\tests;
 
-use BotMan\BotMan\Http\Curl;
-use PHPUnit\Framework\TestCase;
-use BotMan\BotMan\Drivers\NullDriver;
 use BotMan\BotMan\Drivers\DriverManager;
-use BotMan\BotMan\Tests\Fixtures\TestDriver;
+use BotMan\BotMan\Drivers\NullDriver;
+use BotMan\BotMan\Http\Curl;
 use BotMan\BotMan\Tests\Fixtures\AnotherDriver;
+use BotMan\BotMan\Tests\Fixtures\TestDriver;
 use BotMan\BotMan\Tests\Fixtures\TestDriverWithSubDriver;
+use PHPUnit\Framework\TestCase;
 
 class DriverManagerTest extends TestCase
 {
-    protected function tearDown()
+    protected function tearDown(): void
     {
         DriverManager::unloadDriver(TestDriver::class);
         DriverManager::unloadDriver(TestDriverWithSubDriver::class);
