@@ -2,9 +2,10 @@
 
 namespace BotMan\BotMan\Tests\Fixtures;
 
-use BotMan\BotMan\Messages\Incoming\Answer;
+use BotMan\BotMan\Messages\Attachments\Contact;
 use BotMan\BotMan\Messages\Attachments\Location;
 use BotMan\BotMan\Messages\Conversations\Conversation;
+use BotMan\BotMan\Messages\Incoming\Answer;
 
 class TestDataConversation extends Conversation
 {
@@ -53,6 +54,16 @@ class TestDataConversation extends Conversation
                     $this->say($location->getLatitude().':'.$location->getLongitude());
                 }, function (Answer $answer) {
                     $this->say('That is not a location...');
+                });
+            } elseif ($answer->getText() === 'contact') {
+                $this->askForContact('Please supply a contact', function (Contact $contact) {
+                    $this->say($contact->getPhoneNumber().':'.$contact->getFirstName().':'.$contact->getLastName().':'.$contact->getUserId());
+                });
+            } elseif ($answer->getText() === 'custom_contact_repeat') {
+                $this->askForContact('Please supply a contact', function (Contact $contact) {
+                    $this->say($contact->getPhoneNumber().':'.$contact->getFirstName().':'.$contact->getLastName().':'.$contact->getUserId());
+                }, function (Answer $answer) {
+                    $this->say('That is not a contact...');
                 });
             }
         });
