@@ -2,11 +2,11 @@
 
 namespace BotMan\BotMan\Tests\Fixtures;
 
-use BotMan\BotMan\Users\User;
-use BotMan\BotMan\Messages\Incoming\Answer;
 use BotMan\BotMan\Interfaces\DriverInterface;
-use BotMan\BotMan\Messages\Outgoing\Question;
+use BotMan\BotMan\Messages\Incoming\Answer;
 use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use BotMan\BotMan\Messages\Outgoing\Question;
+use BotMan\BotMan\Users\User;
 
 class TestDriverWithSubDriver implements DriverInterface
 {
