@@ -2,8 +2,8 @@
 
 namespace BotMan\BotMan\tests;
 
-use PHPUnit\Framework\TestCase;
 use BotMan\BotMan\Messages\Incoming\IncomingMessage;
+use PHPUnit\Framework\TestCase;
 
 class MessageTest extends TestCase
 {
