@@ -2,10 +2,10 @@
 
 namespace BotMan\BotMan\tests;
 
+use BotMan\BotMan\Messages\Outgoing\Actions\Button;
+use BotMan\BotMan\Messages\Outgoing\Question;
 use Illuminate\Support\Arr;
 use PHPUnit\Framework\TestCase;
-use BotMan\BotMan\Messages\Outgoing\Question;
-use BotMan\BotMan\Messages\Outgoing\Actions\Button;
 
 class QuestionTest extends TestCase
 {
