@@ -1,67 +1,40 @@
 <?php
 
-namespace BotMan\BotMan\Tests;
+namespace BotMan\BotMan\Tests\RedisStorage;
 
+use BotMan\BotMan\Storages\Drivers\RedisStorage;
+use Illuminate\Support\Collection;
+use PHPUnit\Framework\TestCase;
 use Redis;
 use RedisException;
-use PHPUnit\Framework\TestCase;
-use Illuminate\Support\Collection;
-use BotMan\BotMan\Storages\Drivers\RedisStorage;
 
 /**
  * @group integration
+ * @group redis-insecure
  */
-class RedisStorageTest extends TestCase
+class InsecureTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         if (! extension_loaded('redis')) {
             $this->markTestSkipped('Redis extension required');
         }
     }
 
-    public function tearDown()
+    protected function tearDown(): void
     {
         $script = sprintf("for i, name in ipairs(redis.call('KEYS', '%s*')) do redis.call('DEL', name); end", RedisStorage::KEY_PREFIX);
 
         $redis = new Redis();
-        $redis->connect('127.0.0.1');
+        $redis->connect($this->getRedisHost(), $this->getRedisPort());
         $redis->eval($script);
         $redis->close();
-
-        $redis = new Redis();
-        $redis->connect('127.0.0.1', 6380);
-        $redis->auth('secret');
-        $redis->eval($script);
-        $redis->close();
-    }
-
-    /** @test */
-    public function valid_auth()
-    {
-        $storage = new RedisStorage('127.0.0.1', 6380, 'secret');
-        $key = 'key';
-        $data = ['foo' => 1, 'bar' => new \DateTime()];
-        $storage->save($data, $key);
-        self::assertEquals(Collection::make($data), $storage->get($key));
-    }
-
-    /**
-     * @test
-     * @expectedException RedisException
-     */
-    public function invalid_auth()
-    {
-        $storage = new RedisStorage('127.0.0.1', 6380, 'invalid');
-        $key = 'key';
-        $data = ['foo' => 1, 'bar' => new \DateTime()];
-        $storage->save($data, $key);
     }
 
     /** @test */
     public function get()
     {
-        $storage = new RedisStorage();
+        $storage = new RedisStorage($this->getRedisHost(), $this->getRedisPort());
         $key = 'key';
         $data = ['foo' => 1, 'bar' => new \DateTime()];
         $storage->save($data, $key);
@@ -71,7 +44,7 @@ class RedisStorageTest extends TestCase
     /** @test */
     public function delete()
     {
-        $storage = new RedisStorage();
+        $storage = new RedisStorage($this->getRedisHost(), $this->getRedisPort());
         $key = 'key';
         $data = ['foo' => 1, 'bar' => new \DateTime()];
         $storage->save($data, $key);
@@ -84,7 +57,7 @@ class RedisStorageTest extends TestCase
     /** @test */
     public function get_all()
     {
-        $storage = new RedisStorage();
+        $storage = new RedisStorage($this->getRedisHost(), $this->getRedisPort());
         $key1 = 'key1';
         $data1 = ['foo' => 1, 'bar' => new \DateTime()];
         $storage->save($data1, $key1);
@@ -98,5 +71,25 @@ class RedisStorageTest extends TestCase
         self::assertCount(2, $items);
         self::assertEquals(Collection::make($data1), $items[$key1]);
         self::assertEquals(Collection::make($data2), $items[$key2]);
+    }
+
+    /**
+     * Get redis host.
+     *
+     * @return string
+     */
+    protected function getRedisHost()
+    {
+        return $_ENV['REDIS_HOST'] ?? '127.0.0.1';
+    }
+
+    /**
+     * Get redis port.
+     *
+     * @return int
+     */
+    protected function getRedisPort()
+    {
+        return (int) ($_ENV['REDIS_PORT'] ?? 6379);
     }
 }
